Support `rubocop.formatAutocorrectsAll` command
This PR supports `rubocop.formatAutocorrectsAll` command, which requires RuboCop 1.56+.

In addition to the built-in VS Code Formatting API.
It triggers the extension to format and autocorrect all the current file listing by running
the command "RuboCop: Format All with Autocorrects". This is equivalent to `rubocop -A`:

![Autocorrect all command](/docs/autocorrect-all-command.png)

User can use two autocorrect commands depending on the purpose.